requirements.txt
langchain_core
langchain_community
streamlit
langserve
fastapi
uvicorn
sse_starlette
langchain_openai
python-dotenv